Add widget na tela de configuracao Add nova imagem Add codigo para mudar o tamanho da bola e cor pela tela de configuracao

+++ touch_speed/lib/models/configurations.dart
@@ -1,6 +1,6 @@
 class ConfigurationItens {
   final int initialTime;
-  final int ballSize;
+  final double ballSize;
   final bool changeColor;
 
   ConfigurationItens(this.initialTime, this.ballSize, this.changeColor);
